refactor(migrationReducer): Migrating function to reducer
